Updated Guarded API calls

+++ app/api/endpoints/event.py
@@ -1,25 +1,15 @@
 from typing import Annotated
 
-from fastapi import APIRouter, Body
+from fastapi import APIRouter, Body, HTTPException, status, Depends
 
+from app.api.endpoints.users import get_current_user
+from app.schemas.user import User, UserType
 from app.models.event import event_model
 from app.schemas.event import CreateEventRequest, Event, UpdateEventStatusRequestDTO
 
 router = APIRouter()
 
-
-@router.post("/new", response_model=Event)
-async def create_event(event: Annotated[CreateEventRequest, Body(...)]) -> Event:
-    return await event_model.create_event(event)
-
-
-@router.put("/{event_id}", response_model=Event | None)
-async def update_event_status(
-    event_id: str, event: Annotated[UpdateEventStatusRequestDTO, Body(...)]
-) -> Event | None:
-    updated_event = await event_model.update_event_status(event_id, event)
-    return updated_event
-
+# GET
 
 @router.get("/all", response_model=list[Event])
 async def get_events() -> list[Event]:
@@ -37,9 +27,69 @@ async def get_events_by_org(organization_id: str) -> list[Event]:
     event_list = await event_model.get_events_by_organization(organization_id)
     return event_list
 
+# POST
+
+@router.post("/new", response_model=Event)
+async def create_event(
+    event: Annotated[CreateEventRequest, Body(...)],
+    current_user: Annotated[User, Depends(get_current_user)],
+) -> Event:
+    if current_user.user_type not in [UserType.ORGANIZATION, UserType.ADMIN]:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Only users with organization role can create an event",
+        )
+
+    if current_user.entity_id is None:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="You must be associated with an organization to create an event",
+        )
+
+    return await event_model.create_event(event)
+
+# PUT
+
+@router.put("/{event_id}", response_model=Event | None)
+async def update_event_status(
+    event_id: str,
+    event: Annotated[UpdateEventStatusRequestDTO, Body(...)],
+    current_user: Annotated[User, Depends(get_current_user)],
+) -> Event | None:
+
+    if current_user.user_type not in [UserType.ORGANIZATION, UserType.ADMIN]:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Only users with organization role can create an event",
+        )
+
+    if current_user.entity_id is None:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="You must be associated with an organization to create an event",
+        )
+
+    updated_event = await event_model.update_event_status(event_id, event)
+    return updated_event
+
+# DELETE
 
 @router.delete("/{event_id}", response_model=None)
-async def clear_event_by_id(event_id: str) -> None:
+async def clear_event_by_id(
+    event_id: str, current_user: Annotated[User, Depends(get_current_user)]
+) -> None:
+    if current_user.user_type not in [UserType.ORGANIZATION, UserType.ADMIN]:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Only users with organization role can delete an event",
+        )
+
+    if current_user.entity_id is None:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="You must be associated with an organization to delete an event",
+        )
+
     return await event_model.delete_event_by_id(event_id)
